+++ src/features/advanced/index.ts
@@ -4,6 +4,7 @@ import './bbl-collapsible-categories';
 import './bbl-hide-book-value';
 import './bs-hide-zero-workforce';
 import './cogcpex-clean-labels';
+import './context-controls-no-hover';
 import './cxos-hide-exchange';
 import './cxpo-shorten-fields';
 import './finla-hide-ecd';