+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+- `context-controls-no-hover`: New feature. Prevents the context controls from displaying description while hovering over
 - `prod-hide-percent`: New feature. Hides percent value from production lines.
 
 ### Fixed